Backing image
Use a backing image to render from. The idea is that we maintain the current state of the map in there (although I’m not yet keeping it updated) and just copy from it to the screen.

It turns out to not be as fast-a-path as our tiles solution, even though the bitmap format is the same!

In order to draw the changes into it, we need to keep the CGContextRef around and try to draw into it, so we need the NSBitmapImageRep to use the same memory. Maybe using the code in https://searchcode.com/codesearch/view/8166226/ to create an NSBitmapImageRep from the same data buffer that the CGContextRef is using.

I couldn’t get this to work so nicely.